// src/lib/types/enums.ts
export enum UserRole {
  SUPER_ADMIN = "SUPER_ADMIN",
  ADMIN = "ADMIN",
  BOARD = "BOARD",
  DEVELOPER = "DEVELOPER",
  BLOXXS_TEAM = "BLOXXS_TEAM",
  ORIGINADOR = "ORIGINADOR",
  DISTRIBUIDOR = "DISTRIBUIDOR",
  ASSESSOR = "ASSESSOR",
  INVESTIDOR = "INVESTIDOR",
}

export enum InviteStatus {
  PENDENTE = "PENDENTE",
  ACEITO = "ACEITO",
  EXPIRADO = "EXPIRADO",
}